pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ospo-tools"
version = "0.1.0"
description = "A set of tools to help with the OSPO activities and reporting"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Damian Vicino", email = "damian.vicino@datadoghq.com" }
]
urls = { "Homepage" = "https://github.com/DataDog/ospo-tools" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "agithub",
    "scancode-toolkit",
    "typer"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "mypy",
    "black"
]

[project.scripts]
generate_3rd_party_csv = "ospo_tools.generate_3rd_party_csv:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=xml --cov-fail-under=80"

[tool.mypy]
strict = "True"
disallow_untyped_calls = "True"
disallow_untyped_defs = "True"
disallow_incomplete_defs = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
warn_redundant_casts = "True"

[[tool.mypy.overrides]]
module = "scancode"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "agithub.GitHub"
ignore_missing_imports = true
